Rename nested link extension file
