+++ packages/core/lib/nestedLinkAsPlainText.ts
